requirements.txt
markdown-it-py==3.0.0
mdurl==0.1.2
MouseInfo==0.1.3
packaging==25.0
pillow==11.2.1
PyAutoGUI==0.9.54
pyfzf==0.3.1
PyGetWindow==0.0.9
Pygments==2.19.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytesseract==0.3.13
python3-xlib==0.15
pytweening==1.2.0
rich==14.0.0
